Refactor utility and management functions and test for unpacking members data
- enable management command to pass filename to utility function
- enable utility function to take filename, generate filepath and extract JSON from file
- improve utility function handling of null values
- update management command test by isolating from utility command – previously ran utility
and checked whether utility returned a certain snippet. The snippet is now mocked, so
that test does not rely on utility.
- update utility test to check for output snippet that had been checked for in management function
test.

Discovered that there are 21 tests but django counts 20 tests when test suite run.
Looped through all the tests individually and they all ran. It seems like django is
clustering under the hood. Something to explore later.